requirements.txt
asgiref==3.3.4
autopep8==1.5.7
Django==3.2
django-debug-toolbar==3.2.1
pycodestyle==2.7.0
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
typing-extensions==3.7.4.3
